+++ lib/CustomerProductDetail/Model/customer_product_detail_model.dart
@@ -25,7 +25,7 @@ class Product {
   String? productName;
   String? brandName;
   String? shoeType;
-  int? price;
+  double? price;
   String? details;
   List<String>? imagesArray;
   List<String>? sizeArray;
@@ -51,7 +51,7 @@ class Product {
     productName = json['productName'];
     brandName = json['brandName'];
     shoeType = json['shoeType'];
-    price = json['price'];
+    price = json['price'].toDouble();
     details = json['details'];
     imagesArray = json['imagesArray'].cast<String>();
     sizeArray = json['sizeArray'].cast<String>();
